pyproject.toml
[project]
name = "watkit_API"
version = "0.2.0"
description = "API to automate customer service and manage clients"
authors = [
    {name = "Wilson23MLPro", email = "wilsoncarlosraul@gmail.com"}
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "sqlalchemy>=2.0.27",
    "aiosqlite>=0.19.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
    "alembic>=1.15.0",
    "python-multipart>=0.0.22",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.11"
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]